### 2017-05-18 #264 @whedon
[PRE REVIEW]: Mosquitto
**Submitting author:** ralight (<a href="http://orcid.org/0000-0001-9218-7797">Roger A. Light</a>)
**Repository:** <a href="https://github.com/eclipse/mosquitto" target ="_blank">https://github.com/eclipse/mosquitto</a>
**Version:** v1.4.11

Thanks for submitting your paper to JOSS ralight. The JOSS editor (shown at the top of this issue) will work with you on this issue to find a reviewer for your submission before creating the main review issue.

ralight if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](https://github.com/openjournals/joss/blob/master/docs/reviewers.csv) have already agreed to review for JOSS and may be suitable for this submission.
