### 2017-05-22 #265 @whedon
[REVIEW]: Mosquitto
**Submitting author:** @ralight (<a href="http://orcid.org/0000-0001-9218-7797">Roger A. Light</a>)
**Repository:** <a href="https://github.com/eclipse/mosquitto" target ="_blank">https://github.com/eclipse/mosquitto</a>
**Version:** v1.4.11

- [x] **Authorship:** Has the submitting author (ralight) made major contributions to the software?
