### 2017-05-22 #265 @brainstorm
Since, coincidentally I've been using this software recently for my experiments with The Things Network, it has been a breeze to validate its functionality and performance. Good piece of software, @ralight 👍

@arfon, I'm done with it, ready to deposit.
