### 2017-05-22 #265 @ralight
Well awesome, that is a nice coincidence and this was a lot more straightforward than I was expecting. Thanks very much.
